feat: NameInputForm code refactoring
